LOGBACK-1240 Log files aren't deleted on startup with hourly rotation
https://github.com/qos-ch/logback/pull/351

+++ logback-android/src/main/java/ch/qos/logback/core/rolling/helper/RollingCalendar.java
@@ -193,7 +193,7 @@ public class RollingCalendar extends GregorianCalendar {
       case TOP_OF_MINUTE:
         return diff / CoreConstants.MILLIS_IN_ONE_MINUTE;
       case TOP_OF_HOUR:
-        return (int) diff / CoreConstants.MILLIS_IN_ONE_HOUR;
+        return diff / CoreConstants.MILLIS_IN_ONE_HOUR;
       case TOP_OF_DAY:
         return diff / CoreConstants.MILLIS_IN_ONE_DAY;
       case TOP_OF_WEEK:
